+++ src/test/java/com/cinchapi/runway/RecordTest.java
@@ -575,7 +575,7 @@ public class RecordTest extends ClientServerTest {
                 Assert.fail();
             }
             catch (RuntimeException e) {
-                Assert.assertEquals("name must be unique", e.getMessage());
+                Assert.assertTrue(e.getMessage().startsWith("name must be unique"));
             }
 
         }
@@ -781,6 +781,20 @@ public class RecordTest extends ClientServerTest {
                 ImmutableMap.of("name", "b", "email", "b@b.com"));
         Set<?> actual = Collections.ensureSet((Collection<?>) data.get("users"));
         Assert.assertEquals(expected, actual);
+    }
+    
+    @Test
+    public void testSetValueToNullRemovesFromDatabase() {
+        Mock mock = new Mock();
+        mock.alive = true;
+        mock.age = 10;
+        mock.name = "Mock";
+        mock.bar = false;
+        mock.save();
+        mock.age = null;
+        mock.save();
+        mock = runway.load(Mock.class, mock.id());
+        Assert.assertNull(mock.age);
     }
 
     class Node extends Record {